拖动进度view的初步展示

+++ app/src/main/java/com/damon43/polycloudmusic/theinterface/OnBottomPlayStateListener.kt
@@ -4,5 +4,5 @@ package com.damon43.polycloudmusic.theinterface
  * Created by lenovo on 2017/8/29.
  */
 interface OnBottomPlayStateListener {
-    fun onChangeDegree(alpha:Int)
+    fun onChangeDegree(alpha:Float)
 }
